# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Storm Trade Bot API — trading bot for Storm Trade perpetual DEX on TON
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/connect:
    post:
      operationId: connectWallet
      tags: [wallet]
      summary: Connect TON wallet via TonConnect payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletConnectInput"
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletInfo"
...
  /positions/{id}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get a single position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /positions/{id}/close:
    post:
      operationId: closePosition
      tags: [positions]
      summary: Manually close a position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # ── Orders ────────────────────────────────────────────────────────────────
...
  /orders/{id}:
    delete:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel a pending order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # ── Trade History ─────────────────────────────────────────────────────────
...
  /market/price:
    get:
      operationId: getMarketPrice
      tags: [market]
      summary: Get current price for a pair
      parameters:
        - name: pair
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Price
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceData"
  /market/ohlcv:
    get:
      operationId: getOhlcv
      tags: [market]
      summary: Get OHLCV candle data
      parameters:
        - name: pair
          in: query
          required: true
          schema:
            type: string
        - name: interval
          in: query
          required: true
          schema:
            type: string
            enum: [1m, 5m, 15m, 1h, 4h, 1d]
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: OHLCV candles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"

  # ── Price Action & SMC Analysis ──────────────────────────────────────────
  /analysis/levels:
    get:
      operationId: getKeyLevels
      tags: [analysis]
      summary: Get detected key support/resistance levels for a pair
      parameters:
        - name: pair
          in: query
          required: true
          schema:
            type: string
        - name: interval
          in: query
          schema:
            type: string
            default: 1h
      responses:
        "200":
          description: Key levels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyLevelsResult"
  /analysis/zones:
    get:
      operationId: getSupplyDemandZones
      tags: [analysis]
      summary: Get detected supply and demand zones
      parameters:
        - name: pair
          in: query
          required: true
          schema:
            type: string
        - name: interval
          in: query
          schema:
            type: string
            default: 1h
      responses:
        "200":
          description: Supply and demand zones
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZonesResult"
  /analysis/structure:
    get:
      operationId: getMarketStructure
      tags: [analysis]
      summary: Get market structure (HH/HL/LL/LH, BOS, CHoCH) for SMC
      parameters:
        - name: pair
          in: query
          required: true
          schema:
            type: string
        - name: interval
          in: query
          schema:
            type: string
            default: 1h
      responses:
        "200":
          description: Market structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketStructureResult"
  /analysis/signal:
    get:
      operationId: getCurrentSignal
      tags: [analysis]
      summary: Get current algorithm signal and reasoning
      parameters:
        - name: pair
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingSignal"

  # ── Performance Metrics ──────────────────────────────────────────────────
...
components:
  schemas:
...
    ActionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    # ── Wallet ──────────────────────────────────────────────────────────────
    WalletInfo:
      type: object
      required: [connected]
      properties:
        connected:
          type: boolean
        address:
          type: ["string", "null"]
        balance:
          type: ["number", "null"]
        balanceUsd:
          type: ["number", "null"]
        network:
          type: ["string", "null"]
    WalletConnectInput:
      type: object
      required: [address]
      properties:
        address:
          type: string
        proof:
          type: ["string", "null"]

    # ── Risk ────────────────────────────────────────────────────────────────
...
    Position:
      type: object
      required: [id, pair, side, entryPrice, currentPrice, size, sizeUsd, leverage, unrealizedPnlUsd, unrealizedPnlPercent, stopLoss, takeProfit, openedAt, entryReason]
      properties:
        id:
          type: string
        pair:
          type: string
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        currentPrice:
          type: number
        size:
          type: number
        sizeUsd:
          type: number
        leverage:
          type: number
        unrealizedPnlUsd:
          type: number
        unrealizedPnlPercent:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        openedAt:
          type: string
        entryReason:
          type: string
          description: Human-readable reason the algorithm entered

    # ── Orders ──────────────────────────────────────────────────────────────
...
    PriceData:
      type: object
      required: [pair, price, bid, ask, change24h, volume24h, timestamp]
      properties:
        pair:
          type: string
        price:
          type: number
        bid:
          type: number
        ask:
          type: number
        change24h:
          type: number
        volume24h:
          type: number
        timestamp:
          type: string
    Candle:
      type: object
      required: [timestamp, open, high, low, close, volume]
      properties:
        timestamp:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number

    # ── Analysis ─────────────────────────────────────────────────────────────
    KeyLevel:
      type: object
      required: [price, type, strength, touchCount, lastTestedAt]
      properties:
        price:
          type: number
        type:
          type: string
          enum: [support, resistance, pivot]
        strength:
          type: number
          description: Level strength score 0-100
        touchCount:
          type: integer
        lastTestedAt:
          type: string
    KeyLevelsResult:
      type: object
      required: [pair, interval, levels, generatedAt]
      properties:
        pair:
          type: string
        interval:
          type: string
        levels:
          type: array
          items:
            $ref: "#/components/schemas/KeyLevel"
        generatedAt:
          type: string
    Zone:
      type: object
      required: [id, type, priceHigh, priceLow, strength, tested, mitigated, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [supply, demand]
        priceHigh:
          type: number
        priceLow:
          type: number
        strength:
          type: number
        tested:
          type: boolean
        mitigated:
          type: boolean
        createdAt:
          type: string
    ZonesResult:
      type: object
      required: [pair, interval, zones, generatedAt]
      properties:
        pair:
          type: string
        interval:
          type: string
        zones:
          type: array
          items:
            $ref: "#/components/schemas/Zone"
        generatedAt:
          type: string
    StructurePoint:
      type: object
      required: [type, price, timestamp]
      properties:
        type:
          type: string
          enum: [HH, HL, LH, LL, BOS, CHOCH, EQUAL_HIGH, EQUAL_LOW]
        price:
          type: number
        timestamp:
          type: string
    MarketStructureResult:
      type: object
      required: [pair, interval, trend, points, internalTrend, generatedAt]
      properties:
        pair:
          type: string
        interval:
          type: string
        trend:
          type: string
          enum: [bullish, bearish, ranging]
        internalTrend:
          type: string
          enum: [bullish, bearish, ranging]
        points:
          type: array
          items:
            $ref: "#/components/schemas/StructurePoint"
        generatedAt:
          type: string
    TradingSignal:
      type: object
      required: [pair, signal, confluenceScore, reasoning, keyFactors, suggestedEntry, suggestedStopLoss, suggestedTakeProfit, riskReward, generatedAt]
      properties:
        pair:
          type: string
        signal:
          type: string
          enum: [strong_long, long, neutral, short, strong_short]
        confluenceScore:
          type: number
        reasoning:
          type: string
        keyFactors:
          type: array
          items:
            type: string
        suggestedEntry:
          type: ["number", "null"]
        suggestedStopLoss:
          type: ["number", "null"]
        suggestedTakeProfit:
          type: ["number", "null"]
        riskReward:
          type: ["number", "null"]
        generatedAt:
          type: string

    # ── Metrics ──────────────────────────────────────────────────────────────
...